refactor(resource-keys): update hook to leverage ResourceKeys enum
Update hook to use enum in order to aid future developers by preventing mistakes whereproperties are added/removed from Resource interface and the developer is unaware.  To be more specific, without using the enums as a reference point, strings were being used.  It's easy to forget to update those string literals, and thus, run into a hard to find bug.