docs: fix Getting Started > Remix imports
This PR fixes the imports in the Remix Getting Started guide. As far as I know, the `... as React` is used when we import all named exports using `*`, e.g.:

```ts
import * as React from 'react';
```

Instead, here, previously what we did was:

```ts
import React, { createContext } as React from "react"
```

This causes syntax errors in the IDE. However, if this is a valid syntax, please let me know and I'll close this PR. Thanks!