Update number-index.mdx | TypeScript throws error
TypeScript throws errors if we use `isReadOnly: true` in the `getInputProps()` function. Although, if we pass `isReadOnly` does work, there seems to be a type error. However, when using `getInputProps({ readOnly: true });` both TypeScript and functionalities work. So, there is not a change. You just have a stricter type system.

Maybe, change the initial types of the `getInputProps()` parameters?