Add more examples of different Input Methods
The documentation page just shows examples on text and password as input. It would be a great addition to show more examples such as date, month, time and color picker inputs. I saw reddits and stackoverflow questions asking for date and time picker exclusively not being available in Chakra UI as Material UI and ANTD.