refactor: Change tab content id for TOC operation
When the IntersectionObserver finds the DOM to move, the id 'usage'
duplicates to prevent accurate scroll movement. Because the div tag
enclosing the usage tab is using 'usage' id, when 'Usage' is selected
in TOC, the scrolling does not move accurately to the h2#usage tag.
Add the '-wrapper' suffix to the id of the tab content defined in the
content layer file, determining that either id needs to be changed.
Change it to be consistent because other TABs may have overlapping
IDs and causing scrolling problems. As a result, scrolling normally
moves when the 'Usage' menu is clicked in TOC.