refactor: initializing headerRef in cleanUp
Because the H tags are stored in the headerRef, the value is maintained
when the page is moved. If an element with the isIntersecting value of
the IntersectionObserverEntry is true is stored on the previous page,
the topElement cannot be found normally on the next page. Therefore,
when cleaning up useEffect, add a logic to initialize headerRef so that
there is no entry left with isIntersecting true.