refactor: add extra margin at the end of example
Because the way activeId is set has changed, additional space is
required at the bottom of each component's Usage example. For
example, in the case of a Slider component, there is not enough space
at the end of the Usage, so the useSlider at the bottom does not
highlight normally in the TOC. Assign a margin-top value sufficiently
in view-height units to ensure that TOC is activated normally.