Update nextjs-app-guide.mdx (#1679)
Importing the className attribute from the source applies it to the entire site. This way, if we import more than one source, the entire site will be seen with the last one imported.

For this reason, the variable attribute must be imported that allows us to use more than one font on the site using the variable in the theme.ts file as shown at the end of the example.